Agentic_AI/Pipeline_Monitor/requirements.txt
google-generativeai>=0.7.0
streamlit
pandas
sqlalchemy
python-dotenv
prefect>=2.9